requirements.txt
future
envparse==0.2.0
multicolorcaptcha
umongo==3.0.0
faker
ffmpeg-python
humanize
wget==3.2
secureme==0.0.4
cowpy
motor==2.5.1
py-tgcalls==0.8.1
gpytranslate==1.3.0
PyDictionary==2.0.1
emoji
pretty-errors==1.2.25
lxml
textblob==0.17.1
beautifulsoup4
requests
sqlalchemy==1.3.23
python-telegram-bot==13.11
cloudscraper==1.2.58
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
psutil
jikanpy
speedtest-cli==2.1.3
coffeehouse
pymongo==3.12.0
regex
bleach
ptbcontrib @ git+https://github.com/python-telegram-bot/ptbcontrib.git@907070d129b7cf80b0236d72fb498667b5b22b63
git+https://github.com/trentm/python-markdown2
wikipedia
telethon==1.24.0
spamwatch
alphabet_detector
cachetools
dnspython==2.1.0
fontTools
gtts
tswift
pytz
PyYAML>=5.3.1
redis==3.5.3
aioredis==2.0.0
hiredis==2.0.0
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
instantmusic
PyLyrics
hachoir
pybase64
pySmartDL
validators
pyrogram==1.4.0
asyncio
nltk
aiohttp>=3.7.4
pykeyboard==0.1.2
Pillow==9.0.0
CurrencyConverter
googletrans==4.0.0-rc1
google_trans_new
pyrate-limiter
nekos.py
tgcrypto
python-arq==6.0.3
hachoir
configparser
python-dotenv==0.19.2
bing_image_downloader
parsel
urllib3
heroku3
telegraph==2.1.0
youtube_search
search_engine_parser
glitch_this
NumPy
opencv-python-headless
countryinfo==0.1.2
python-rapidjson==1.5
pytimeparse
AnilistPython
pyjokes
lyricsgenius
lottie
wand
pyrogram
tgcrypto
requests
aiofiles>=0.6.0
aiohttp[speedups]>=3.7.3
tracemoepy
lottie
bs4
motor
dnspython
natsort
apscheduler
lxml
html5lib
pyfy
